[#7] Service 로직 구현 및 단위 테스트 작성

+++ api/src/main/java/com/lightningbid/api/item/dto/request/ItemPatchRequestDto.java
@@ -2,6 +2,7 @@ package com.lightningbid.api.item.dto.request;
 
 import lombok.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -17,6 +18,6 @@ public class ItemPatchRequestDto {
     private List<String> imageIds;
     private Boolean isDirectTrade;
     private String location;
-    private Integer startPrice;
+    private BigDecimal startPrice;
     private LocalDateTime auctionEndTime;
 }
